feat: core watch files
