+++ .changeset/soft-flies-run.md
@@ -1,0 +1,7 @@
+---
+'@modern-js/core': minor
+'@modern-js/app-tools': minor
+---
+
+feat: `modern-js/core` can watch files change
+feat: `modern-js/core` 可以监听文件变化
